requirements.txt
pandas
tqdm
scikit-learn
python-dotenv
google-generativeai